requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.0.0
openai==1.30.0
python-dotenv==1.0.0
requests==2.33.0
